requirements.txt
numpy==1.26.4
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
simpleaudio==1.0.4
sounddevice==0.5.1
torch==2.6.0
transformers==4.49.0
langchain
langchain-openai
langchain-core
langchain-community
pydantic
openai
langgraph
python-dotenv
tavily-python